+++ packages/mod_quizdashboard/mod_quizdashboard.php
@@ -3,7 +3,7 @@
 /**
  * @package     QuizKit
  * @subpackage  mod_quizdashboard
- * @version     1.0.0
+ * @version     1.1.1
  * @author      Michelle Ermen
  * @copyright   Copyright © 2023 MSE Digital All Rights Reserved
  * @license     GNU GPLv3 <http://www.gnu.org/licenses/gpl.html> or later; see LICENSE.txt
@@ -21,7 +21,7 @@ ToolBarHelper::title(Text::_('Quiz Dashboard'), 'quiz-dashboard');
 
 $db = Factory::getDbo();
 $query = $db->getQuery(true);
-$query->select($db->quoteName(array('id', 'email', 'params', 'score', 'submission_time')));
+$query->select($db->quoteName(array('id', 'email', 'params', 'score', 'visitor_id', 'submission_time')));
 $query->from($db->quoteName('#__quizkit_submissions'));
 $db->setQuery($query);
 $results = $db->loadObjectList();