+++ script.install.php
@@ -2,7 +2,7 @@
 
 /**
  * @package     QuizKit
- * @version     1.1.0
+ * @version     1.1.1
  * @author      Michelle Ermen
  * @copyright   Copyright © 2023 MSE Digital All Rights Reserved
  * @license     GNU GPLv3 <http://www.gnu.org/licenses/gpl.html> or later; see LICENSE.txt
@@ -34,17 +34,19 @@ class pkg_QuizKitInstallerScript extends InstallerScript
     $query = "CREATE TABLE IF NOT EXISTS #__quizkit_submissions (
       id int(11) NOT NULL AUTO_INCREMENT,
       email varchar(50) NOT NULL,
-      params varchar(255) NOT NULL,
+      params TEXT NOT NULL,
+      visitor_id int(11) NOT NULL DEFAULT 0,
       score float NOT NULL,
       submission_time datetime NOT NULL,
       PRIMARY KEY (id)
     )";
 
     $db->setQuery($query);
-    $result = $db->execute();
-
-    if (!$result) {
-      Factory::getApplication()->enqueueMessage($db->stderr(), 'error');
+    try {
+      $db->execute();
+      echo '<p>The module has been installed.</p>';
+    } catch (Exception $e) {
+      Error::raiseWarning(500, $e->getMessage());
       return false;
     }
     echo '<p>The module has been installed.</p>';
@@ -60,40 +62,13 @@ class pkg_QuizKitInstallerScript extends InstallerScript
   {
     $db = Factory::getDBO();
 
-    // Verwijder de tabel
-    $query = "DROP TABLE IF EXISTS #__quizkit_submissions";
-    $db->setQuery($query);
-    $result = $db->execute();
-
-    if (!$result) {
-      Factory::getApplication()->enqueueMessage($db->stderr(), 'error');
+    try {
+      $db->execute();
+      echo '<p>The module has been uninstalled and the table has been removed.</p>';
+    } catch (Exception $e) {
+      Error::raiseWarning(500, $e->getMessage());
       return false;
     }
-
-    // Verwijder de bestanden van de module
-    jimport('joomla.filesystem.folder');
-    jimport('joomla.filesystem.file');
-
-    $modulePaths = [
-      JPATH_SITE . '/modules/mod_quizmaker',
-      JPATH_ADMINISTRATOR . '/modules/mod_quizdashboard'
-    ];
-
-    foreach ($modulePaths as $path) {
-      if (JFolder::exists($path)) {
-        JFolder::delete($path);
-      }
-    }
-
-    // Verwijder de vermeldingen uit de `extensions` tabel
-    $query = $db->getQuery(true)
-      ->delete($db->quoteName('#__extensions'))
-      ->where($db->quoteName('element') . ' = ' . $db->quote('mod_quizmaker'))
-      ->orWhere($db->quoteName('element') . ' = ' . $db->quote('mod_quizdashboard'));
-    $db->setQuery($query);
-    $db->execute();
-
-    echo '<p>The module has been uninstalled and the table has been removed.</p>';
   }
 
   /**
@@ -123,25 +98,36 @@ class pkg_QuizKitInstallerScript extends InstallerScript
       $db->execute();
     }
 
-    // Create the table if it does not exist
+    // set default to 0 on column visitor_id
+    $query = 'ALTER TABLE ' . $db->quoteName('#__quizkit_submissions') . ' MODIFY COLUMN ' . $db->quoteName('visitor_id') . ' int(11) NOT NULL DEFAULT 0';
+    $db->setQuery($query);
+    $db->execute();
+
     $query = "CREATE TABLE IF NOT EXISTS #__quizkit_submissions (
-            id int(11) NOT NULL AUTO_INCREMENT,
-            email varchar(50) NOT NULL,
-            params TEXT NOT NULL,
-            score float NOT NULL,
-            submission_time datetime NOT NULL,
-            PRIMARY KEY (id)
-        )";
+        id int(11) NOT NULL AUTO_INCREMENT,
+        email varchar(50) NOT NULL,
+        params TEXT NOT NULL,
+        visitor_id int(11) NOT NULL DEFAULT 0,
+        score float NOT NULL,
+        submission_time datetime NOT NULL,
+        PRIMARY KEY (id)
+    )";
 
     $db->setQuery($query);
-    $result = $db->execute();
 
-    if (!$result) {
-      Factory::getApplication()->enqueueMessage($db->stderr(), 'error');
+    $db->setQuery($query);
+    try {
+      $db->execute();
+
+      // Load the manifest file to get the version
+      $manifest = $parent->getParent()->manifest;
+      $version = (string) $manifest->version;
+
+      echo '<p>The module has been updated to version ' . $version . '.</p>';
+    } catch (Exception $e) {
+      Error::raiseWarning(500, $e->getMessage());
       return false;
     }
-
-    echo '<p>The module has been updated to version ' . $parent->getManifest()->version . '.</p>';
   }
 
   /**