+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Change log
 
-## v1.1.0
-- Overall Joomla 5 compatibility
+## v1.1.1
+- Update column visitor_id to default to 0
+
+## v1.1.0 
+- Update column 'params' from varchar to type 'text'
+- replace 'query' function to 'execute' function
 
 ## v1.0.3
 - (mod_quizmaker/quizmaker.js): change button[type="submit"] to specific selector for submit button to prevent bleeding